+++ protoc_plugin/test/golden_file.dart
@@ -13,7 +13,8 @@ import 'package:test/test.dart';
 void expectMatchesGoldenFile(String actual, String goldenFilePath) {
   File goldenFile = File(goldenFilePath);
   if (goldenFile.existsSync()) {
-    expect(actual, equals(goldenFile.readAsStringSync()));
+    expect(actual, equals(goldenFile.readAsStringSync()),
+        reason: 'goldenFilePath: "${goldenFilePath}"');
   } else {
     // This enables writing the updated file when the run in otherwise hermetic
     // settings.